requirements.txt
scikit-learn
numpy
Pillow
joblib
flask
flask-cors
gunicorn
